NewFix

+++ database/migrations/2025_02_22_153702_create_products_table.php
@@ -12,8 +12,8 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->text('description')->nullable();
-            $table->decimal('price', 10, 2);
-            $table->decimal('price_sale', 15, 2);
+            $table->integer('price');
+            $table->integer('price_sale');
             $table->string('image')->nullable();
             // $table->integer('stock');
             $table->string('product_detail')->nullable();
