+++ database/migrations/2025_03_30_212848_create_new_table.php
@@ -13,13 +13,13 @@ return new class extends Migration
     {
         Schema::create('payments', function (Blueprint $table) {
             $table->id();
-    $table->foreignId('order_id')->constrained('orders')->onDelete('cascade');
-    $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
-    $table->string('payment_method');
-    $table->decimal('amount', 10, 2);
-    $table->string('status')->default('pending');
-    $table->string('transaction_id')->unique();
-    $table->timestamps();
+            $table->foreignId('order_id')->constrained('orders')->onDelete('cascade');
+            $table->foreignId('user_id')->constrained('users')->onDelete('cascade');
+            $table->string('payment_method');
+            $table->integer('amount');
+            $table->string('status')->default('pending');
+            $table->string('transaction_id')->unique();
+            $table->timestamps();
         });
     }
